+++ src/Controller/RegistrationController.php
@@ -23,12 +23,25 @@ class RegistrationController extends AbstractController
 
     private VerifyEmailHelperInterface $verifyEmailHelper;
     private MailerInterface $mailer;
-    private CheckUserService $checkUser;
 
     public function __construct(VerifyEmailHelperInterface $helper, MailerInterface $mailer)
     {
         $this->verifyEmailHelper = $helper;
         $this->mailer = $mailer;
+    }
+
+    function checkData(User $user): array
+    {
+        # Errors existent Nickname and/or Email
+        $errors = [];
+        $repo = $this->getDoctrine()->getRepository(User::class);
+        if ($repo->count(['username' => $user->getUsername()]) > 0) {
+            $errors['nick'] = "This nickname already exists";
+        }
+        if ($repo->count(['email' => $user->getEmail()]) > 0) {
+            $errors['email'] = "This e-mail address already exists";
+        }
+        return $errors;
     }
 
     /**
@@ -40,7 +53,7 @@ class RegistrationController extends AbstractController
         $form = $this->createForm(RegisterType::class, $user);
         $form->handleRequest($request);
         if ($form->isSubmitted() && !$form->isValid()) {
-            $errors = $this->checkUser->checkData($user);
+            $errors = $this->checkData($user);
             if (!empty($errors)) {
                 return $this->render('registration/register.html.twig', [
                     'errors' => $errors,
@@ -50,7 +63,7 @@ class RegistrationController extends AbstractController
             }
         }
         if ($form->isSubmitted() && $form->isValid()) {
-            $errors = $this->checkUser->checkData($user);
+            $errors = $this->checkData($user);
             if (!empty($errors)) {
                 return $this->render('registration/register.html.twig', [
                     'errors' => $errors,
